+++ tests/unit/Dumper/Config/DumperConfigTest.php
@@ -43,7 +43,7 @@ class DumperConfigTest extends TestCase
                 'table1' => ['truncate' => true],
                 'table2' => ['limit' => 1],
                 'table3' => ['orderBy' => 'field1'],
-                'table4' => ['converters' => ['field1' => 'randomizeEmail']],
+                'table4' => ['converters' => ['field1' => ['converter' => 'randomizeEmail']]],
             ],
         ];